refactor: compress into layout
