+++ app/theming.js
@@ -1,29 +1,14 @@
 "use client"
 
-import Dropdown from "@components/dropdown"
-import Icon from "@components/icon"
 import classNames from "@utils/classnames"
 import { useTheme } from "next-themes"
 import { useEffect, useState } from "react"
 
 const themes = ["autumn", "pastel", "coffee", "emerald", "bumblebee", "luxury", "cyberpunk", "cupcake", "synthwave", "corporate", "retro", "valentine", "night", "halloween", "dracula", "forest"]
-
-/**
- * color theme picker
- */
-export default function ThemePicker() {
-  return (
-    <Dropdown className="dropdown-bottom dropdown-end">
-      <Dropdown.Button children={<Icon shape="M4.098 19.902a3.75 3.75 0 005.304 0l6.401-6.402M6.75 21A3.75 3.75 0 013 17.25V4.125C3 3.504 3.504 3 4.125 3h5.25c.621 0 1.125.504 1.125 1.125v4.072M6.75 21a3.75 3.75 0 003.75-3.75V8.197M6.75 21h13.125c.621 0 1.125-.504 1.125-1.125v-5.25c0-.621-.504-1.125-1.125-1.125h-4.072M10.5 8.197l2.88-2.88c.438-.439 1.15-.439 1.59 0l3.712 3.713c.44.44.44 1.152 0 1.59l-2.879 2.88M6.75 17.25h.008v.008H6.75v-.008z" />} />
-      <Dropdown.Menu className="max-h-64 overflow-y-auto dropdown-content scrollbar-hide" children={<Themes />} />
-    </Dropdown>
-  )
-}
-
 /**
  * theme picker menu
  */
-function Themes() {
+export default function Themes() {
   const { setTheme, theme } = useTheme()
   const [mounted, setMounted] = useState(false)
